# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Grimoire API
  version: "0.1.0"
  description: |
    Event-sourced TTRPG campaign manager.
    All write routes emit domain events; reads return the current aggregate snapshot.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CharacterKind:
      type: string
      enum: [PlayerCharacter, NonPlayer]
    CharacterStatus:
      type: string
      enum: [Active, Departed, Deceased]
...
    Character:
      type: object
      properties:
        id:         { type: string, format: uuid }
        story_id:   { type: string, format: uuid }
        name:       { type: string }
        backstory:  { type: string }
        kind:       { $ref: "#/components/schemas/CharacterKind" }
        status:     { $ref: "#/components/schemas/CharacterStatus" }
        metadata:   { $ref: "#/components/schemas/Metadata" }
...
    Metadata:
      type: object
      properties:
        owner: { type: string, format: uuid }
paths:
...
  /player/characters/{id}:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    get:
      summary: Get own character
      tags: [Player — Characters]
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
        "404": { $ref: "#/components/responses/NotFound" }
  /player/characters/{id}/name:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    patch:
      summary: Rename own character
      tags: [Player — Characters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "204": { description: Updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /player/characters/{id}/backstory:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    patch:
      summary: Update own character backstory
      tags: [Player — Characters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [backstory]
              properties:
                backstory: { type: string }
      responses:
        "204": { description: Updated }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Shared response components ─────────────────────────────────────────────
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation error (empty name, blank title, etc.)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...